LAI/requirements.txt
pyttsx3
SpeechRecognition
pygame
google-generativeai
